requirements.txt
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
requests==2.32.5
peewee==3.15.4